Backend/requirements.txt
asgiref==3.8.1
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.10
distlib==0.3.9
Django==5.2.3
djangorestframework==3.16.0
Eel==0.9.0
filelock==3.18.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
platformdirs==4.3.7
playsound==1.2.2
psycopg2==2.9.10
pycparser==2.22
pyjokes==0.8.3
pypiwin32==223
pyttsx3==2.98
pytz==2025.2
pywin32==308
pyzbar==0.1.9
qrcode==8.2
requests==2.32.3
scipy==1.15.3
setuptools==75.8.0
sounddevice==0.5.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.30.0
virtualenvwrapper-win==1.2.7
wavio==0.0.9
zope.event==5.0
zope.interface==7.2
